Move timeOfDayInSeconds computation to ShipsClockFramework
